+++ NEWS.md
@@ -1,5 +1,9 @@
 # openair (development version)
 
+## Bug Fixes
+
+-  `timeAverage` now has a more robust approach to multi-period averaging time such as "3 day". Bin boundaries are at fixed periods of time, which should ensure consistency across data sets that start at different times. This also means there is less need to use the argument `start.date` unless there is a need to extend the start date of the data for some reason e.g. to the beginning of a year. THis change may result in slight differences in returned output but should not affect periods such as "day", "hour" and "month".
+
 # openair 3.0.0
 
 ## Dependency Changes